Move dlisio.open to it's own file
