+++ python/dlisio/__init__.py
@@ -3,7 +3,8 @@ from . import plumbing
 from . import errors
 from .settings import get_encodings, set_encodings
 from .file import physicalfile, logicalfile
-from .load import open, load
+from .load import load
+from .open import open
 
 try:
     import pkg_resources
