+++ python/dlisio/load.py
@@ -2,29 +2,9 @@ import os
 
 from . import core
 from .file import physicalfile, logicalfile
+from .open import open
 from .errors import ErrorHandler
 
-def open(path, offset = 0):
-    """ Open a file
-
-    Open a low-level file handle. This is not intended for end-users - rather,
-    it's an escape hatch for very broken files that dlisio cannot handle.
-
-    Parameters
-    ----------
-    path : str_like
-    offset: int
-        Physical file offset at which handle must be opened
-
-    Returns
-    -------
-    stream : dlisio.core.stream
-
-    See Also
-    --------
-    dlisio.load
-    """
-    return core.open(str(path), offset)
 
 def load(path, error_handler = None):
     """ Loads a file and returns one filehandle pr logical file.
